+++ novawallet/Modules/Staking/Validation/StakingDataValidatorFactory.swift
@@ -9,7 +9,7 @@ struct MinStakeIsNotViolatedParams {
     let votersCount: UInt32?
 }
 
-protocol StakingDataValidatingFactoryProtocol: BaseDataValidatingFactoryProtocol {
+protocol StakingDataValidatingFactoryProtocol: StakingBaseDataValidatingFactoryProtocol {
     func canUnbond(amount: Decimal?, bonded: Decimal?, locale: Locale) -> DataValidating
     func canRebond(amount: Decimal?, unbonding: Decimal?, locale: Locale) -> DataValidating
 
@@ -70,27 +70,17 @@ protocol StakingDataValidatingFactoryProtocol: BaseDataValidatingFactoryProtocol
     ) -> DataValidating
 
     func allowsNewNominators(flag: Bool, locale: Locale) -> DataValidating
-
-    func minStakeNotCrossed(
-        for inputAmount: Decimal,
-        params: MinStakeCrossedParams,
-        chainAsset: ChainAsset,
-        locale: Locale
-    ) -> DataValidating
 }
 
-final class StakingDataValidatingFactory {
-    weak var view: (ControllerBackedProtocol & Localizable)?
-
-    var basePresentable: BaseErrorPresentable { presentable }
-
-    let presentable: StakingErrorPresentable
-
-    let balanceFactory: BalanceViewModelFactoryProtocol?
+final class StakingDataValidatingFactory: StakingBaseDataValidatingFactory {
+    private let presentable: StakingErrorPresentable
 
     init(presentable: StakingErrorPresentable, balanceFactory: BalanceViewModelFactoryProtocol? = nil) {
         self.presentable = presentable
-        self.balanceFactory = balanceFactory
+        super.init(
+            presentable: presentable,
+            balanceFactory: balanceFactory
+        )
     }
 }
 
@@ -216,58 +206,6 @@ extension StakingDataValidatingFactory: StakingDataValidatingFactoryProtocol {
             } else {
                 return false
             }
-        })
-    }
-
-    func minStakeNotCrossed(
-        for inputAmount: Decimal,
-        params: MinStakeCrossedParams,
-        chainAsset: ChainAsset,
-        locale: Locale
-    ) -> DataValidating {
-        let inputAmountInPlank = inputAmount.toSubstrateAmount(
-            precision: chainAsset.assetDisplayInfo.assetPrecision
-        ) ?? 0
-
-        let stakedAmountInPlank = params.stakedAmountInPlank
-        let minStake = params.minStake
-
-        return WarningConditionViolation(onWarning: { [weak self] delegate in
-            guard let balanceFactory = self?.balanceFactory else {
-                return
-            }
-
-            let stakedAmount = stakedAmountInPlank ?? 0
-            let diff = stakedAmount >= inputAmountInPlank ? stakedAmount - inputAmountInPlank : 0
-
-            let minStakeDecimal = (minStake ?? 0).decimal(precision: chainAsset.asset.precision)
-            let diffDecimal = diff.decimal(precision: chainAsset.asset.precision)
-
-            let minStakeString = balanceFactory.amountFromValue(minStakeDecimal).value(for: locale)
-            let diffString = balanceFactory.amountFromValue(diffDecimal).value(for: locale)
-
-            self?.presentable.presentCrossedMinStake(
-                from: self?.view,
-                minStake: minStakeString,
-                remaining: diffString,
-                action: {
-                    params.unstakeAllHandler()
-                    delegate.didCompleteWarningHandling()
-                },
-                locale: locale
-            )
-
-        }, preservesCondition: {
-            guard
-                let stakedAmountInPlank = stakedAmountInPlank,
-                let minStake = minStake,
-                stakedAmountInPlank >= inputAmountInPlank else {
-                return false
-            }
-
-            let diff = stakedAmountInPlank - inputAmountInPlank
-
-            return diff == 0 || diff >= minStake
         })
     }
